# SimpleGroceryListApplication/simple-grocery-list/GroceryListAPISpecs.yml
openapi: 3.0.2
info:
  title: Simple Grocery List
  version: 1.0.0
  description: This application is to view, add, and delete items from a grocery list.
  contact:
    name: Casey Fleck
    url: https://github.com/CaseyJoFleck/SimpleGroceryList
    email: casejofleck@gmail.com
servers:
  - url: http://localhost:8080
    description: LOCAL
paths:
  /groceries:
    get:
      tags:
        - Groceries
      summary: Get All Groceries
      description: '## Get all groceries from grocery list'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroceryItem'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Not Found:
                  value:
                    message: No groceries found.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500ErrorResponse'
          description: Internal Server Error
    post:
      summary: Add Item
      description: '## Add Grocery Item'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroceryItem'
        required: true
      tags:
        - Groceries
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryItem'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Invalid Quantity:
                  value:
                    message: The quantity of items must be greater than 0.
                    timestamp: '2022-02-06T19:19:21.9424049'
                Missing Field:
                  value:
                    message: Grocery item name must not be blank when adding new item.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Bad Request
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500ErrorResponse'
          description: Internal Server Error
    delete:
      tags:
        - Groceries
      summary: Remove All Groceries
      description: '## Delete all groceries from list'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Delete All Groceries:
                  value:
                    message: All grocery items have been removed from list.
                    timestamp: '2022-02-06T19:07:32.3406995'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Not Found:
                  value:
                    message: No groceries found.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Internal Server Error:
                  value:
                    message: An internal error has occurred. Please try again later.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Internal Server Error
  /groceries/{id}:
    put:
      tags:
        - Groceries
      summary: Add Quantity
      description: '## Add Quantity'
      parameters:
        - examples:
            Item Id:
              value: '3'
          name: id
          description: Grocery item id.
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryItem'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Not Found:
                  value:
                    message: Operation cannot continue. Item not found in grocery list.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500ErrorResponse'
          description: Internal Server Error
    delete:
      tags:
        - Groceries
      operationId: RemoveOneItemQuantity
      summary: Remove one Quantity
      description: '## Decrease item quantity by one. If only one item exists, item will be removed item from list'
      parameters:
        - examples:
            Item Id:
              value: '3'
          name: id
          description: Grocery item id.
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Ok:
                  value:
                    message: Grocery item Eggs has been removed from list.
                    timestamp: '2022-02-06T19:07:32.3406995'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Not Found:
                  value:
                    message: Operation cannot continue. Item not found in grocery list.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500ErrorResponse'
          description: Internal Server Error
  /groceries/all/{id}:
    delete:
      tags:
        - Groceries
      summary: Remove Grocery
      description: '## Remove Grocery By Id'
      parameters:
        - examples:
            id:
              value: '5'
          name: id
          description: id of item to delete
          schema:
            type: number
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Deleted:
                  value:
                    message: Bread has been removed from list.
                    timestamp: '2022-02-06T19:07:32.3406995'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBody'
              examples:
                Not Found:
                  value:
                    message: Operation cannot continue. Item not found in grocery list.
                    timestamp: '2022-02-06T19:19:21.9424049'
          description: Not Found
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500ErrorResponse'
          description: Internal Server Error
components:
  schemas:
    GroceryItem:
      title: GroceryItem
      description: An item from a grocery list.
      required:
        - name
        - quantity
      type: object
      properties:
        id:
          format: int32
          description: Unique id related to item.
          type: integer
          example: '5'
        name:
          description: The item name.
          type: string
          example: Butter
        quantity:
          description: The amount of the item.
          type: integer
          example: '2'
    AddGroceryItem:
      title: AddGroceryItem
      description: A grocery item to add.
      required:
        - name
        - quantity
      type: object
      properties:
        name:
          description: The item name.
          type: string
          example: Butter
        quantity:
          description: The amount of the item.
          type: integer
          example: '2'
    ResponseBody:
      title: ErrorResponseNotFound
      description: Groceries were not found
      type: object
      properties:
        message:
          description: Message of body.
          type: string
        timestamp:
          description: Timestamp in IS0 format.
          type: string
    500ErrorResponse:
      title: 500ErrorResponse
      description: Internal Server Error
      type: object
      properties:
        message:
          description: Message of error.
          type: string
          example: "An internal error has occurred. Please try again later."
        timestamp:
          description: Timestamp in ISO format.
          type: string
          example: "2022-02-06T19:19:21.9424049"